Refactor more utilities
