+++ clowder/fork.py
@@ -6,11 +6,11 @@ import os
 
 from termcolor import cprint
 
-from clowder.utility.clowder_utilities import existing_git_repository
 from clowder.utility.git.printing import (
     project_ref_string,
     project_string
 )
+
 
 class Fork(object):
     """clowder.yaml fork class"""
@@ -32,7 +32,7 @@ class Fork(object):
 
     def print_status(self):
         """Print formatted fork status"""
-        if not existing_git_repository(self.path):
+        if not GitRepo.existing_git_repository(self.path):
             cprint(self.path, 'green')
             return
         project_output = project_string(self.path, self.path)
